#118 Replace remote calls to WireMock stub calls for power operations test. changes according comments
