+++ sdk/src/test/java/com/centurylink/cloud/sdk/server/services/dsl/servers/operations/ServerActionTest.java
@@ -12,9 +12,29 @@ import static com.centurylink.cloud.sdk.tests.TestGroups.RECORDED;
  * @author Aliaksandr Krasitski
  */
 @Test(groups = RECORDED)
-public class ServerRestoreTest extends AbstractServerOperationTest implements WireMockMixin {
+public class ServerActionTest extends AbstractServerOperationTest implements WireMockMixin {
 
     Group group = Group.refById("68c30c5898584105902306f7f610b31b");
+
+    @Test
+    @WireMockFileSource("/archive")
+    public void testArchiveServer() {
+        serverService.archive(server)
+            .waitUntilComplete();
+
+        ServerMetadata serverMetadata = serverService.findByRef(server);
+        assertThatServerHasStatus(serverMetadata, ARCHIVED);
+    }
+
+    @Test
+    @WireMockFileSource("/reboot")
+    public void testRebootServer() {
+        serverService.reboot(server)
+            .waitUntilComplete();
+
+        ServerMetadata serverMetadata = serverService.findByRef(server);
+        assertThatServerHasStatus(serverMetadata, ACTIVE);
+    }
 
     @Test
     @WireMockFileSource("/restore")
@@ -26,5 +46,4 @@ public class ServerRestoreTest extends AbstractServerOperationTest implements Wi
         assertThatServerHasStatus(serverMetadata, ACTIVE);
     }
 }
-
