+++ sdk/src/test/java/com/centurylink/cloud/sdk/server/services/dsl/servers/operations/ServerSnapshotTest.java
@@ -11,7 +11,27 @@ import static com.centurylink.cloud.sdk.tests.TestGroups.RECORDED;
  * @author Aliaksandr Krasitski
  */
 @Test(groups = RECORDED)
-public class ServerRevertToSnapshotTest extends AbstractServerOperationTest implements WireMockMixin {
+public class ServerSnapshotTest extends AbstractServerOperationTest implements WireMockMixin {
+
+    @Test
+    @WireMockFileSource("/snapshot/create")
+    public void testCreateSnapshotServer() {
+        serverService.createSnapshot(server)
+            .waitUntilComplete();
+
+        ServerMetadata serverMetadata = serverService.findByRef(server);
+        assert serverMetadata.getDetails().getSnapshots().size() == 1;
+    }
+
+    @Test
+    @WireMockFileSource("/snapshot/delete")
+    public void testDeleteSnapshotServer() {
+        serverService.deleteSnapshot(server)
+            .waitUntilComplete();
+
+        ServerMetadata serverMetadata = serverService.findByRef(server);
+        assert serverMetadata.getDetails().getSnapshots().size() == 0;
+    }
 
     @Test
     @WireMockFileSource("/snapshot/revert")
